+++ src/main/java/com/smartsheet/api/models/enums/SeatType.java
@@ -17,5 +17,5 @@
 package com.smartsheet.api.models.enums;
 
 public enum SeatType {
-    MEMBER, PROVISIONAL_MEMBER, GUEST, VIEWER
+    MEMBER, PROVISIONAL_MEMBER, GUEST, VIEWER, CONTRIBUTOR
 }