+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - WireMock integration tests for contract testing for POST /2.0/reports/{id}/scope and DELETE /2.0/reports/{id}/scope endpoints
 - Support for the POST /2.0/reports endpoint
 - WireMock integration tests for contract testing for POST /2.0/reports endpoint
+- Added CONTRIBUTOR to SeatType enum
+- Added CONTRIBUTOR to DowngradeSeatType enum
+- Support for `displayContributorSeatType` query parameter in `listUsers` and `listUserPlans` methods
+- Added WireMock integration tests for CONTRIBUTOR seat type in list users and downgrade user endpoints
 
 ## [3.10.1] - 2026-03-31
 ### Fixed